Add GPGSingleExtensionMode configuration option
Currently, vim-gnupg implicitly expects the user to operate on
filesnames of the format:

file.txt.gpg

When user saves/reads the file, respective autocommands are issued
for the stripped filename:

file.txt (.gpg extension stripped)

However, this might not always be the desired behaviour - indeed, user
might explicitly want to run autommands on the gnupg recognized
extension, i.e. file.gpg.

This commit adds a new configuration variable GPGSingleExtensionMode
which is set to 0 by default. If set to 1, vim-gnupg will not strip the
file extensions when issuing autocommands.